+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE='cardiolink-v3.10.0';
+const CACHE='cardiolink-v3.10.1';
 const SHELL=['./','./index.html','./styles.css?v=3100','./app.js?v=3100','./manifest.webmanifest?v=3100','./icons/icon-192.png','./icons/icon-512.png'];
 self.addEventListener('install',e=>{e.waitUntil(caches.open(CACHE).then(c=>c.addAll(SHELL)).then(()=>self.skipWaiting()))});
 self.addEventListener('activate',e=>{e.waitUntil(caches.keys().then(keys=>Promise.all(keys.filter(k=>k!==CACHE).map(k=>caches.delete(k)))).then(()=>self.clients.claim()))});